Fix compile warnings on Linux
git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@1949 d8f3d347-7192-4e35-a67b-6d701703d080

+++ dtnetsubs.c
@@ -556,7 +556,7 @@ const char *dtNowString (void)
     gettimeofday (&tv, NULL);
     strftime (ts, sizeof (ts) - 1, "%y/%m/%d %H.%M.%S.", 
               localtime_r (&tv.tv_sec, &tmbuf));
-    sprintf (us, "%06d.", tv.tv_usec);
+    sprintf (us, "%06d.", (int)tv.tv_usec);
 #endif
     strcat (ts, us);
     return ts;
